+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   [[#34]](https://github.com/Marcin-Migdal/m-component-library/issues/34) **[PATCH]** Fix clear icon in dropdown wrong color
 -   [[#35]](https://github.com/Marcin-Migdal/m-component-library/issues/35) **[PATCH]** Add noBottomMargin props to all inputs that on default have bottom margin
 -   [[#33]](https://github.com/Marcin-Migdal/m-component-library/issues/33) **[PATCH]** Add readOnly prop to inputs
+-   [[#55]](https://github.com/Marcin-Migdal/m-component-library/issues/55) **[PATCH]** Add classNamesObj to input components
 
 ### 28.09-2024 - Release 1.0.3