+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### ? - Release 1.0.4
+### 18.10.2024 - Release 1.0.4
 
 -   [[#37]](https://github.com/Marcin-Migdal/m-component-library/issues/37) **[PATCH]** Alert close only when overlay click started on overlay
 -   [[#36]](https://github.com/Marcin-Migdal/m-component-library/issues/36) **[PATCH]** Add hslToHex helper function